# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Linktic",
        "description": "Documentación de API para empresas y empleados",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/employees": {
            "get": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Obtiene los registros de empleados activos",
                "operationId": "2058499f556b4bb9477e07e7a0912c43",
                "responses": {
                    "200": {
                        "description": "Obtiene las empresas activas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Company 1"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "company1@example.com"
                                                    },
                                                    "logo": {
                                                        "type": "string",
                                                        "example": "company1_logo.png"
                                                    },
                                                    "website": {
                                                        "type": "string",
                                                        "example": "https://www.company1.com"
                                                    },
                                                    "sttaus": {
                                                        "type": "boolean",
                                                        "example": "true"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-04-20T17:30:22"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2023-02-23 12:33:45"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Crea un nuevo empleado",
                "operationId": "7683a7b1aec5103190be815e1e30f7da",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nombre de la empresa"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "empresa@example.com"
                                    },
                                    "logo": {
                                        "description": "Logo de la empresa (archivo)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "website": {
                                        "type": "string",
                                        "example": "https://www.empresa.com"
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": "true"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Empresa creada exitosamente"
                    },
                    "419": {
                        "description": "Token CSRF mismatch"
                    }
                }
            }
        },
        "/api/employees/{id}": {
            "put": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Actualiza los datos de una empresa",
                "operationId": "5cc6df4bd0c5be4084900d70ecc36a25",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del empleado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nuevo nombre de la empresa"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "nuevoemail@empresa.com"
                                    },
                                    "website": {
                                        "type": "string",
                                        "example": "https://www.nuevaempresa.com"
                                    },
                                    "logo": {
                                        "description": "Nuevo logo de la empresa (archivo)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Empresa actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Nuevo nombre de la empresa"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email",
                                            "example": "nuevoemail@empresa.com"
                                        },
                                        "logo": {
                                            "type": "string",
                                            "example": "nuevo-nombre-de-la-empresa-2024-04-20-23-31-05.png"
                                        },
                                        "website": {
                                            "type": "string",
                                            "example": "https://www.nuevaempresa.com"
                                        },
                                        "status": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-04-20T23:31:05"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-04-21T00:45:15"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Elimina una empresa",
                "description": "No elimina el registro solo cambia de activo a inactivo, esto para no perder información, desde el Front-end se puede manejar este estado para mostrar o no mostrar los registros",
                "operationId": "bd97927cf2dedf50b48ebf1a3bfc487e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del empleado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Empresa eliminada correctamente"
                    }
                }
            }
        }
    }
}